+++ lizard_portal/media/lizard_portal/vss/model/Esf.js
@@ -12,9 +12,12 @@ Ext.define('Vss.model.Esf', {
             {name: 'config_id', type: 'auto'},
             {name: 'name', type: 'string'},
             {name: 'source_name', type: 'auto'},
-            {name: 'manual_value', type: 'float', default_value: -1},
-            {name: 'manual', type: 'int', default_value: -1},
-            {name: 'auto_value', type: 'string', default_value: '-'},
-            {name: 'type', type: 'string'}/*,
+            {name: 'manual_value', type: 'auto'},
+            {name: 'auto_value', type: 'auto'},
+            {name: 'manual', type: 'int'},
+            {name: 'is_manual', type: 'boolean'},
+            {name: 'type', type: 'string'},
+            {name: 'config_type', type: 'string'}
+            /*,
             {name: 'qtip', mapping: 'last_comment', type: 'string'}*/]
 });