+++ lizard_portal/templates/portals/ekr-score.js
@@ -1,18 +1,18 @@
 {
-    itemId: 'maatregelen',
-    title: 'Maatregelen',
+    itemId: 'ekr-score',
+    title: 'EKR-score',
 	xtype: 'portalpanel',
     breadcrumbs: [
     {
         name: 'KRW-overzicht',
         link: 'krw-overzicht'
     },{
-        name: 'Maatregelen'
+        name: 'EKR-score'
     }],
 	items:[{
 		flex: 1,
 		items: [{
-			title: 'Maatregelen',
+			title: 'EKR-score',
             flex:1
 		}]
 	}]