+++ lizard_portal/templates/portals/krw-overzicht.js
@@ -5,6 +5,8 @@
  * Time: 17:52
  * To change this template use File | Settings | File Templates.
  */
+{% load get_portal_template %}
+
 {
     itemId: 'krw-overzicht',
     title: 'KRW overzicht',
@@ -15,68 +17,10 @@
     }],
 	items:[{
 		width: 300,
-		items: [{
-			title: 'Gebiedsinformatie',
-            flex:1,
-            plugins: [
-                'applycontext'
-            ],
-            autoScroll: true,
-            loader: {
-                ajaxOptions: {
-                    method: 'GET'
-                },
-                loadMask: true,
-                url: '/portal/configuration/',
-                autoLoad: false,
-                 baseParams: {
-                     _accept: 'text/html',
-                     portalTemplate: 'eigenschappen'
-                 }
-            },
-            //xtype: "image",
-            applyParams: function(params) {
-                 var me = this;
-                 me.getLoader().load({
-                     url: '/portal/configuration/',
-                     params: {
-                         object_id: params.object_id
-                     }
-                 });
-            }
-
-		},{
-			title: 'Communique',
-            bodyCls: 'l-grid',
-            height: 150,
-            collapsed: true,
-            autoScroll: true,
-            plugins: [
-                'applycontext'
-            ],
-            loader: {
-                ajaxOptions: {
-                    method: 'GET'
-                },
-                loadMask: true,
-                url: '/portal/configuration/',
-                autoLoad: false,
-                baseParams: {
-                    _accept: 'text/html',
-                    portalTemplate: 'communique'
-                }
-            },
-            applyParams: function(params) {
-                var me = this;
-                me.getLoader().load({
-                    url: '/portal/configuration/',
-                    params: {
-                        object_id: params.object_id
-                    }
-                });
-
-            }
-        },{
+		items: [
+            {% get_portal_template gebiedseigenschappen %},
+            {% get_portal_template communique %},
+        {
 			title: 'Extra info?',
             flex:1
 		}]
@@ -86,37 +30,9 @@
             title: 'Grafieken',
             flex: 1,
             xtype: 'multigraph',
-            graph_service_url: '/map/adapter/adapter_fewsnorm/image/',
-            adapter_layer_json: {module_id:null,parameter_id:"ALMR110","fews_norm_source_slug":""},
-            graphs: [{
-                title: 'Stuurparameter 1',
-                timeseries:[{
-                    parameter_id: "ALMR110",
-                    module_id: "ImportLE",
-                    ident: "53R0017"
-                }]
-            },{
-                title: 'Stuurparameter 2',
-                timeseries:[{
-                    parameter_id: "ALMR110",
-                    module_id: "ImportLE",
-                    ident: "53R0017"
-                }]
-            },{
-                title: 'EKR scores',
-                timeseries:[{
-                    parameter_id: "ALMR110",
-                    module_id: "ImportLE",
-                    ident: "53R0017"
-                }]
-            }, {
-                title: 'Maatregel voortgang',
-                timeseries:[{
-                    parameter_id: "ALMR110",
-                    module_id: "ImportLE",
-                    ident: "53R0017"
-                }]
-            }]
+            graph_service_url: '/graph/',
+            context_manager: Ext.getCmp('portalWindow').context_manager,
+            graphs: {% get_portal_template graphs-krw-overzicht %}
 		}]
 	}]
 }